+++ R/MsBackendTimsTof-functions.R
@@ -23,13 +23,13 @@
         list(frames, indices)
     }, BPPARAM = BPPARAM)
     x@frames <- do.call(rbindFill, lapply(L, "[[", 1))
-    idx <- match(colnames(x@frames, .SPECTRA_VARIABLE_MAPPINGS))
+    idx <- match(colnames(x@frames), .SPECTRA_VARIABLE_MAPPINGS)
     not_na <- !is.na(idx)
     colnames(x@frames)[not_na] <- names(.SPECTRA_VARIABLE_MAPPINGS)[idx[not_na]]
     if (any(colnames(x@frames) == "polarity"))
         x@frames$polarity <- .format_polarity(x@frames$polarity)
     x@indices <- do.call(rbind, lapply(L, "[[", 2))
-    x@fileNames <- file
+    x@fileNames <- setNames(seq_len(length(file)), file)
     x
 }
 
@@ -44,14 +44,14 @@
 #' @noRd
 #'
 .valid_required_columns <- function(x, columns = character(0)) {
-  if (nrow(x)) {
-    missing_cn <- setdiff(columns, colnames(x))
-    if (length(missing_cn))
-      return(paste0("Required column(s): ",
-                    paste(missing_cn, collapse = ", "),
-                    " is/are missing"))
-  }
-  NULL
+    if (nrow(x)) {
+        missing_cn <- setdiff(columns, colnames(x))
+        if (length(missing_cn))
+            return(paste0("Required column(s): ",
+                          paste(missing_cn, collapse = ", "),
+                          " is/are missing"))
+    }
+    NULL
 }
 
 #' @description
@@ -63,7 +63,7 @@
 #'
 #' @noRd
 .valid_frames <- function(x) {
-  .valid_required_columns(x, c("Id", "file"))
+    .valid_required_columns(x, c("frameId", "file"))
 }
 
 #' @description
@@ -74,30 +74,33 @@
 #'
 #' @noRd
 .valid_indices <- function(x) {
-  msg <- .valid_required_columns(x@indices, c("frame", "file"))
-  if ("file" %in% colnames(x@indices) &&
-      any(!x@indices[, "file"] %in% seq_len(length(x@fileNames))))
-    msg <- c(msg, "Some file indices are not valid")
-  if (any(!paste0(x@indices[, "frame"], x@indices[, "file"]) %in%
-          paste0(x@frames$Id, x@frames$file)))
-    msg <- c(msg, "Some indices in x@indices are not compatible with x@frames")
-  msg
+    msg <- .valid_required_columns(x@indices, c("frame", "file"))
+    if ("file" %in% colnames(x@indices) && !setequal(x@indices[, "file"],
+                                                     x@fileNames))
+        msg <- c(msg, "Some file indices are out of bounds")
+    if (any(!paste0(x@indices[, "frame"], x@indices[, "file"]) %in%
+            paste0(x@frames$frameId, x@frames$file)))
+        msg <- c(msg, "Some indices are out of bounds")
+    msg
 }
 
 #' @description
 #'
-#' Checks if the `character` `x` is compatible to be the `@fileNames` slot of a
-#' `MsBackendTimsTof` object.
+#' Checks if the named `integer` `x` is compatible to be the `@fileNames` slot
+#' of a `MsBackendTimsTof` object.
 #'
 #' @param x `character` with folder names.
 #'
 #' @noRd
 .valid_fileNames <- function(x) {
-  msg <- NULL
-  if (anyNA(x))
-    msg <- "'NA' values in fileNames are not allowed."
-  msg <- c(msg, Spectra:::.valid_ms_backend_files_exist(unique(x)))
-  msg
+    msg <- NULL
+    if (anyNA(x))
+        msg <- "'NA' values in fileNames names are not allowed."
+    nms <- names(x)
+    if (anyNA(nms))
+        msg <- "'NA' values in fileNames names are not allowed."
+    msg <- c(msg, Spectra:::.valid_ms_backend_files_exist(unique(nms)))
+    msg
 }
 
 #' @importFrom methods new
@@ -106,41 +109,43 @@
 #'
 #' @export
 MsBackendTimsTof <- function() {
-  new("MsBackendTimsTof")
+    new("MsBackendTimsTof")
 }
 
-#' Read variables from the `OpenTIMS` object (`OpenTIMS(x)`) created from data
-#' stored in folder `x`. At least one variable in `OpenTIMS(x)@all_columns`
+#' Get `x@all_columns` variables (including "`mz`" and "`intensity`") from `x`
+#' splitted by spectra. At least one variable among `x@all_columns` and
 #' different from `"frame"` and "`scan`" has to be provided via `columns`
-#' parameter.
+#' parameter. 
 #'
-#' @param x `character(1)` path to the '*.d' folder to read from.
+#' @param x `MsBackendTimsTof` object.
 #'
 #' @param columns `character` with the names of the columns to extract.
-#'
-#' @param indices `matrix` of indices. It contains columns `"frame"` and
-#' `"scan"` to select which rows of data to read from `OpenTIMS(x)`.
-#'
-#' @return `list` of `numeric` (if a single variable is read) or `list`
-#' of `matrix` (otherwise). Each element of the list corresponds to certain
-#' `"frame"` and `"scan"` indeces from `indeces`.
 #'
 #' @importFrom opentimsr OpenTIMS CloseTIMS query
 #'
 #' @noRd
-.read_frame_col <- function(x, columns, indices) {
-  tms <- OpenTIMS(x)
-  on.exit(opentimsr::CloseTIMS(tms))
-  if (any(!columns %in% tms@all_columns))
-    stop("Invalid value for columns")
-  if (!length(sd <- setdiff(columns, c("frame", "scan"))))
-    stop("At least one column value different from 'frame' and 'scan' required")
-  tmp <- query(tms, unique(indices[, "frame"]), c("frame", "scan", sd))
-  f <- factor(paste(tmp$frame, tmp$scan),
-              levels = paste(indices[, "frame"], indices[, "scan"]))
-  if (length(sd) == 1)
-    unname(split(tmp[, sd], f))
-  else unname(split.data.frame(as.matrix(tmp[, sd]), f))
+.get_tims_columns <- function(x, columns) {
+    res <- vector(mode = "list", length(x))
+    nms <- names(x@fileNames)
+    for (i in seq_len(length(nms))) {
+        I <- which(x@indices[, "file"] == x@fileNames[i])
+        tms <- OpenTIMS(nms[i])
+        on.exit(opentimsr::CloseTIMS(tms))
+        if (any(notin <- !columns %in% tms@all_columns))
+            stop("Column(s) ",
+                 paste0("'", columns[notin], "'", collapse = ", "),
+                 " not available.", call. = FALSE)
+        if (!length(sd <- setdiff(columns, c("frame", "scan"))))
+            stop("At least one column value different from 'frame' and",
+                 " 'scan' required")
+        tmp <- query(tms, unique(x@indices[I, "frame"]), c("frame", "scan", sd))
+        f <- factor(paste(tmp$frame, tmp$scan),
+                    levels = paste(x@indices[I, "frame"], x@indices[I, "scan"]))
+        if (length(sd) == 1)
+            res[I] <- unname(split(tmp[, sd], f))
+        else res[I] <- unname(split.data.frame(as.matrix(tmp[, sd]), f))
+    }
+    res
 }
 
 #' Extract columns from @frames given the @indices in `x`. The function takes
@@ -155,11 +160,11 @@ MsBackendTimsTof <- function() {
 #' @noRd
 .get_frame_columns <- function(x, columns) {
     if (!all(columns %in% colnames(x@frames)))
-        stop("Columns ", paste0("'", columns[!columns %in% colnames(x@frames)],
-                                "'", collapse = ", "), " not available.",
-             call. = FALSE)
+        stop("Column(s) ",
+             paste0("'", columns[!columns %in% colnames(x@frames)],
+                    "'", collapse = ", "), " not available.", call. = FALSE)
     idx <- match(paste(x@indices[, "frame"], x@indices[, "file"]),
-                 paste(x@frames$Id, x@frames$file))
+                 paste(x@frames$frameId, x@frames$file))
     x@frames[idx, columns]
 }
 
@@ -168,7 +173,8 @@ MsBackendTimsTof <- function() {
 #' @noRd
 .SPECTRA_VARIABLE_MAPPINGS <- c(
     rtime = "Time",
-    polarity = "Polarity"
+    polarity = "Polarity",
+    frameId = "Id"
 )
 
 .format_polarity <- function(x) {
@@ -176,4 +182,52 @@ MsBackendTimsTof <- function() {
     xn[grep("^(p|\\+)", x, ignore.case = TRUE)] <- 1L
     xn[grep("^(n|-)", x, ignore.case = TRUE)] <- 0L
     xn
+}
+
+# can we assume that tms@all_coulmns is the same for all the TimsTOF? 
+.TIMSTOF_COLUMNS <- c("mz", "intensity", "tof", "inv_ion_mobility")
+
+.spectra_data <- function(x, columns = spectraVariables(x)) {
+    if (!all(present <- columns %in% spectraVariables(x)))
+        stop("Column(s) ", paste0("\"", columns[!present], "\"",
+                                  collapse = ", "),
+             " not available.", call. = FALSE)
+    res <- vector(mode = "list", length = length(columns))
+    names(res) <- columns
+    core_cols <- columns[columns %in% names(Spectra:::.SPECTRA_DATA_COLUMNS)]
+    frames_cols <- columns[columns %in% colnames(x@frames)]
+    tims_cols <- columns[columns %in% .TIMSTOF_COLUMNS]
+    if ("scanIndex" %in% columns) {
+        res[["scanIndex"]] <- x@indices[, "scan"]
+        core_cols <- core_cols[core_cols != "scanIndex"]
+    }
+    if ("frameId" %in% frames_cols) {
+        res[["frameId"]] <- x@indices[, "frame"]
+        frames_cols <- frames_cols[frames_cols != "frameId"]
+    }
+    if ("file" %in% frames_cols) {
+        res[["file"]] <- x@indices[, "file"]
+        frames_cols <- frames_cols[frames_cols != "file"]
+    }
+    if (length(frames_cols)) {
+        res[frames_cols] <- .get_frame_columns(x, frames_cols)
+        core_cols <- setdiff(core_cols, frames_cols)
+    }
+    if (length(tims_cols)) {
+        res[tims_cols] <- lapply(tims_cols, function(col)
+            NumericList(lapply(.get_tims_columns(x, tims_cols),
+                               function(m) unname(m[, col])), compress = FALSE))
+        core_cols <- setdiff(core_cols, tims_cols)
+    }
+    if ("dataStorage" %in% columns) {
+        res[["dataStorage"]] <- dataStorage(x)
+        core_cols <- core_cols[core_cols != "dataStorage"]
+    }
+    if (length(core_cols)) {
+        res[core_cols] <- lapply(Spectra:::.SPECTRA_DATA_COLUMNS[core_cols],
+                                 function(z, n) rep(as(NA, z), n), length(x))
+    }
+    if (length(res))
+        as(res, "DataFrame")
+    else NULL
 }