+++ R/MsBackendTimsTof.R
@@ -44,10 +44,22 @@
 #' - `rtime`: gets the retention times for each spectrum. Returns a `numeric`
 #'   vector (length equal to the number of spectra) with the retention time
 #'   for each spectrum.
+#'   
+#' - `spectraData`: gets spectra variables (specified by `columns`) from
+#'   `object`.
+#'
+#' - `spectraVariables`: returns a `character` vector with the spectra variables
+#'   names of core spectra variables defined in the Spectra package and other
+#'   additional variables contained in `object`. Note that also `"mz"` and
+#'   `"intensity"` (which are by default not returned by the
+#'   `spectraVariables,Spectra` method) are returned.
 #'
 #' @param BPPARAM Parameter object defining the parallel processing
 #' setup to import data in parallel. Defaults to `BPPARAM = bpparam()`.
 #' See [bpparam()] for more information.
+#' 
+#' @param columns For `spectraData`: names of the spectra variables to extract
+#'   from `object`.
 #'
 #' @param drop For `[`: not considered.
 #'
@@ -72,20 +84,24 @@ setClass("MsBackendTimsTof",
          contains = "MsBackend",
          slots = c(frames = "data.frame",
                    indices = "matrix",
-                   fileNames = "character"),
+                   fileNames = "integer"),
          prototype = prototype(frames = data.frame(),
-                               indices = matrix(),
-                               fileNames = character(),
+                               indices = matrix(nrow = 0, ncol = 3,
+                                                dimnames = list(NULL, 
+                                                                c("frame",
+                                                                  "scan",
+                                                                  "file"))),
+                               fileNames = integer(),
                                readonly = TRUE,
                                version = "0.1"))
 
 #' @importFrom methods validObject
 setValidity("MsBackendTimsTof", function(object) {
-  msg <- .valid_fileNames(object@fileNames)
-  msg <- c(msg, .valid_frames(object@frames))
-  msg <- c(msg, .valid_indices(object))
-  if (length(msg)) msg
-  else TRUE
+    msg <- .valid_fileNames(object@fileNames)
+    msg <- c(msg, .valid_frames(object@frames))
+    msg <- c(msg, .valid_indices(object))
+    if (length(msg)) msg
+    else TRUE
 })
 
 #' @importFrom BiocParallel bplapply
@@ -93,33 +109,31 @@ setValidity("MsBackendTimsTof", function(object) {
 #' @rdname MsBackendTimsTof
 setMethod("backendInitialize", signature = "MsBackendTimsTof",
           function(object, files, ..., BPPARAM = bpparam()) {
-            if (missing(files) || !length(files))
-              stop("Parameter 'files' is mandatory for 'MsBackendMzR'")
-            if (!is.character(files))
-              stop("Parameter 'files' is expected to be a character vector",
-                   " with the files names from where data should be",
-                   " imported")
-            files <- normalizePath(files, mustWork = FALSE)
-            msg <- Spectra:::.valid_ms_backend_files_exist(files)
-            if (length(msg))
-              stop(msg)
-            object <- .initialize(object, files, BPPARAM)
-            validObject(object)
-            object
+              if (missing(files) || !length(files))
+                  stop("Parameter 'files' is mandatory for 'MsBackendMzR'")
+              if (!is.character(files))
+                  stop("Parameter 'files' is expected to be a character vector",
+                       " with the files names from where data should be",
+                       " imported")
+              files <- normalizePath(files, mustWork = FALSE)
+              msg <- Spectra:::.valid_ms_backend_files_exist(files)
+              if (length(msg))
+                  stop(msg)
+              object <- .initialize(object, files, BPPARAM)
+              validObject(object)
+              object
           })
 
 #' @exportMethod length
 #'
 #' @rdname MsBackendTimsTof
 setMethod("length", "MsBackendTimsTof", function(x) {
-  nrow(x@indices)
+    nrow(x@indices)
 })
 
 #' @rdname MsBackendTimsTof
 setMethod("peaksData", "MsBackendTimsTof", function(object) {
-  do.call(c, lapply(seq_len(length(object@fileNames)), function(i)
-    .read_frame_col(object@fileNames[i], c("mz", "intensity"),
-                    object@indices[object@indices[, "file"] == i, 1:2])))
+    .get_tims_columns(object, c("mz", "intensity"))
 })
 
 
@@ -127,20 +141,14 @@ setMethod("peaksData", "MsBackendTimsTof", function(object) {
 #'
 #' @rdname MsBackendTimsTof
 setMethod("mz", "MsBackendTimsTof", function(object) {
-  NumericList(do.call(c, lapply(seq_len(length(object@fileNames)), function(i)
-    .read_frame_col(object@fileNames[i], "mz",
-                    object@indices[object@indices[, "file"] == i, 1:2]))),
-    compress = FALSE)
+    NumericList(.get_tims_columns(object, "mz"), compress = FALSE)
 })
 
 #' @importFrom IRanges NumericList
 #'
 #' @rdname MsBackendTimsTof
 setMethod("intensity", "MsBackendTimsTof", function(object) {
-  NumericList(do.call(c, lapply(seq_len(length(object@fileNames)), function(i)
-    .read_frame_col(object@fileNames[i], "intensity",
-                    object@indices[object@indices[, "file"] == i, 1:2]))),
-    compress = FALSE)
+    NumericList(.get_tims_columns(object, "intensity"), compress = FALSE)
 })
 
 #' @rdname MsBackendTimsTof
@@ -154,22 +162,36 @@ setMethod("rtime", "MsBackendTimsTof", function(object) {
 #'
 #' @rdname MsBackendTimsTof
 setMethod("[", "MsBackendTimsTof", function(x, i, j, ..., drop = FALSE) {
-  if (missing(i))
-    return(x)
-  i <- i2index(i, length(x))
-  slot(x, "indices", check = FALSE) <- x@indices[i, , drop = FALSE]
-  ff_indices <- paste(x@indices[, "frame"], x@indices[, "file"])
-  slot(x, "frames", check = FALSE) <-
-    x@frames[match(unique(ff_indices),
-                   paste(x@frames$Id, x@frames$file)), , drop = FALSE]
-  # x@frames$NumScans <- unname(table(ff_indices)) should we update NumScans?
-  slot(x, "fileNames", check = FALSE) <- x@fileNames[unique(x@frames$file)]
-  x
+    if (missing(i))
+        return(x)
+    i <- i2index(i, length(x))
+    slot(x, "indices", check = FALSE) <- x@indices[i, , drop = FALSE]
+    ff_indices <- paste(x@indices[, "frame"], x@indices[, "file"])
+    slot(x, "frames", check = FALSE) <-
+        x@frames[match(unique(ff_indices),
+                       paste(x@frames$frameId, x@frames$file)), , drop = FALSE]
+    slot(x, "fileNames", check = FALSE) <-
+        x@fileNames[x@fileNames %in% unique(x@frames$file)]
+    x
 })
 
 #' @rdname MsBackendTimsTof
 setMethod("dataStorage", "MsBackendTimsTof", function(object) {
-  if("file" %in% colnames(object@indices))
-    return (object@fileNames[object@indices[, "file"]])
-  character(0)
+    if("file" %in% colnames(object@indices) && length(object@fileNames))
+        return (names(object@fileNames[match(object@indices[, "file"],
+                                             object@fileNames)])) 
+    character(0)
 })
+
+#' @rdname MsBackendTimsTof
+setMethod("spectraVariables", "MsBackendTimsTof", function(object) {
+    unique(c(names(Spectra:::.SPECTRA_DATA_COLUMNS), .TIMSTOF_COLUMNS,
+             colnames(object@frames)))
+})
+
+#' @rdname MsBackendTimsTof
+setMethod("spectraData", "MsBackendTimsTof",
+          function(object, columns = spectraVariables(object)) {
+              .spectra_data(object, columns)
+          })
+